Usage of doubleValue method for some CreatureStat cases.

+++ L2J_Mobius_4.0_GrandCrusade/java/org/l2jmobius/gameserver/model/actor/stat/CreatureStat.java
@@ -805,7 +805,7 @@ public class CreatureStat
 		{
 			if (_statsAdd.containsKey(stat))
 			{
-				return _statsAdd.get(stat);
+				return _statsAdd.get(stat).doubleValue();
 			}
 			return defaultValue;
 		}
@@ -836,7 +836,7 @@ public class CreatureStat
 		{
 			if (_statsMul.containsKey(stat))
 			{
-				return _statsMul.get(stat);
+				return _statsMul.get(stat).doubleValue();
 			}
 			return defaultValue;
 		}
@@ -853,7 +853,7 @@ public class CreatureStat
 	 */
 	public double getValue(Stat stat, double baseValue)
 	{
-		return _fixedValue.containsKey(stat) ? _fixedValue.get(stat) : stat.finalize(_creature, OptionalDouble.of(baseValue));
+		return _fixedValue.containsKey(stat) ? _fixedValue.get(stat).doubleValue() : stat.finalize(_creature, OptionalDouble.of(baseValue));
 	}
 	
 	/**
@@ -862,7 +862,7 @@ public class CreatureStat
 	 */
 	public double getValue(Stat stat)
 	{
-		return _fixedValue.containsKey(stat) ? _fixedValue.get(stat) : stat.finalize(_creature, OptionalDouble.empty());
+		return _fixedValue.containsKey(stat) ? _fixedValue.get(stat).doubleValue() : stat.finalize(_creature, OptionalDouble.empty());
 	}
 	
 	protected void resetStats()
@@ -998,11 +998,11 @@ public class CreatureStat
 				changed = new HashSet<>();
 				for (Stat stat : Stat.values())
 				{
-					if ((_statsAdd.containsKey(stat) ? _statsAdd.get(stat) : stat.getResetAddValue()) != (adds.containsKey(stat) ? adds.get(stat) : stat.getResetAddValue()))
+					if ((_statsAdd.containsKey(stat) ? _statsAdd.get(stat).doubleValue() : stat.getResetAddValue()) != (adds.containsKey(stat) ? adds.get(stat).doubleValue() : stat.getResetAddValue()))
 					{
 						changed.add(stat);
 					}
-					else if ((_statsMul.containsKey(stat) ? _statsMul.get(stat) : stat.getResetMulValue()) != (muls.containsKey(stat) ? muls.get(stat) : stat.getResetMulValue()))
+					else if ((_statsMul.containsKey(stat) ? _statsMul.get(stat).doubleValue() : stat.getResetMulValue()) != (muls.containsKey(stat) ? muls.get(stat).doubleValue() : stat.getResetMulValue()))
 					{
 						changed.add(stat);
 					}
@@ -1045,7 +1045,7 @@ public class CreatureStat
 		final Map<Position, Double> map = _positionStats.get(stat);
 		if ((map != null) && map.containsKey(position))
 		{
-			return map.get(position);
+			return map.get(position).doubleValue();
 		}
 		return 1d;
 	}
@@ -1060,7 +1060,7 @@ public class CreatureStat
 		final Map<MoveType, Double> map = _moveTypeStats.get(stat);
 		if ((map != null) && map.containsKey(type))
 		{
-			return map.get(type);
+			return map.get(type).doubleValue();
 		}
 		return 0d;
 	}
@@ -1072,7 +1072,7 @@ public class CreatureStat
 	
 	public double getReuseTypeValue(int magicType)
 	{
-		return _reuseStat.containsKey(magicType) ? _reuseStat.get(magicType) : 1d;
+		return _reuseStat.containsKey(magicType) ? _reuseStat.get(magicType).doubleValue() : 1d;
 	}
 	
 	public void mergeReuseTypeValue(int magicType, double value, BiFunction<? super Double, ? super Double, ? extends Double> func)
@@ -1082,7 +1082,7 @@ public class CreatureStat
 	
 	public double getMpConsumeTypeValue(int magicType)
 	{
-		return _mpConsumeStat.containsKey(magicType) ? _mpConsumeStat.get(magicType) : 1d;
+		return _mpConsumeStat.containsKey(magicType) ? _mpConsumeStat.get(magicType).doubleValue() : 1d;
 	}
 	
 	public void mergeMpConsumeTypeValue(int magicType, double value, BiFunction<? super Double, ? super Double, ? extends Double> func)
@@ -1095,7 +1095,7 @@ public class CreatureStat
 		final LinkedList<Double> skillEvasions = _skillEvasionStat.get(magicType);
 		if ((skillEvasions != null) && !skillEvasions.isEmpty())
 		{
-			return skillEvasions.peekLast();
+			return skillEvasions.peekLast().doubleValue();
 		}
 		return 0d;
 	}
